receiver/requirements.txt
aiokafka==0.8.0
anyio==3.6.2
async-timeout==4.0.2
bidict==0.22.1
certifi==2022.12.7
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.92.0
fonttools==4.38.0
gevent==22.10.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.0
orjson==3.8.6
packaging==23.0
Pillow==9.4.0
pydantic==1.10.5
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.21.1
python-engineio==4.3.4
python-multipart==0.0.5
python-socketio==5.7.2
pytz==2022.7.1
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websocket==0.2.1
websockets==10.4
zope.event==4.6
zope.interface==5.5.2
